refactor(integrations): reuse shared harness utilities
