+++ packages/integrations/core/src/harness/env.ts
@@ -1,8 +1,10 @@
 /** Only STAGEHAND_* and BROWSERBASE_* host env vars cross into harness processes. */
-export function buildAllowlistedEnv(): Record<string, string> {
+export function buildAllowlistedEnv(
+  source: NodeJS.ProcessEnv = process.env,
+): Record<string, string> {
   const env: Record<string, string> = {};
-  for (const [key, value] of Object.entries(process.env)) {
-    if (/^(STAGEHAND_|BROWSERBASE_)/.test(key) && value) {
+  for (const [key, value] of Object.entries(source)) {
+    if (/^(STAGEHAND_|BROWSERBASE_)/u.test(key) && value) {
       env[key] = value;
     }
   }
